refactor p2p test
