+++ p2p/p2p_test.go
@@ -1,15 +1,10 @@
 package p2p
 
 import (
-	"bytes"
 	"encoding/hex"
+	"sync/atomic"
 
-	"github.com/33cn/chain33/common/merkle"
-
-	//"fmt"
-	"net"
 	"os"
-	"sort"
 	"strings"
 	"testing"
 	"time"
@@ -27,17 +22,15 @@ import (
 )
 
 var (
-	q           queue.Queue
-	p2pModule   *P2p
-	dataDir     = "testdata"
 	testChannel = int32(119)
 )
 
 func init() {
-	os.RemoveAll(dataDir)
 	l.SetLogLevel("err")
-	q = queue.New("channel")
-	go q.Start()
+}
+
+func processMsg(q queue.Queue) {
+
 	go func() {
 
 		cfg, sub := types.InitCfg("../cmd/chain33/chain33.test.toml")
@@ -119,10 +112,6 @@ func init() {
 			}
 		}
 	}()
-	time.Sleep(time.Second)
-	p2pModule = newP2p(53802, dataDir, q)
-	p2pModule.Wait()
-
 }
 
 //new p2p
@@ -152,11 +141,9 @@ func freeP2p(p2p *P2p) {
 	if err := os.RemoveAll(p2p.cfg.DbPath); err != nil {
 		log.Error("removeTestDbErr", "err", err)
 	}
-	p2p.client.CloseQueue()
 }
 
-func TestP2PEvent(t *testing.T) {
-	qcli := q.Client()
+func testP2PEvent(t *testing.T, qcli queue.Client) {
 	msg := qcli.NewMessage("p2p", types.EventBlockBroadcast, &types.Block{})
 	qcli.Send(msg, false)
 
@@ -178,17 +165,17 @@ func TestP2PEvent(t *testing.T) {
 	qcli.Send(msg, false)
 
 }
-func TestNetInfo(t *testing.T) {
-	p2pModule.node.nodeInfo.IsNatDone()
-	p2pModule.node.nodeInfo.SetNatDone()
-	p2pModule.node.nodeInfo.Get()
-	p2pModule.node.nodeInfo.Set(p2pModule.node.nodeInfo)
-	assert.NotNil(t, p2pModule.node.nodeInfo.GetListenAddr())
-	assert.NotNil(t, p2pModule.node.nodeInfo.GetExternalAddr())
+func testNetInfo(t *testing.T, p2p *P2p) {
+	p2p.node.nodeInfo.IsNatDone()
+	p2p.node.nodeInfo.SetNatDone()
+	p2p.node.nodeInfo.Get()
+	p2p.node.nodeInfo.Set(p2p.node.nodeInfo)
+	assert.NotNil(t, p2p.node.nodeInfo.GetListenAddr())
+	assert.NotNil(t, p2p.node.nodeInfo.GetExternalAddr())
 }
 
 //测试Peer
-func TestPeer(t *testing.T) {
+func testPeer(t *testing.T, p2p *P2p, q queue.Queue) {
 
 	conn, err := grpc.Dial("localhost:53802", grpc.WithInsecure(),
 		grpc.WithDefaultCallOptions(grpc.UseCompressor("gzip")))
@@ -198,7 +185,7 @@ func TestPeer(t *testing.T) {
 	remote, err := NewNetAddressString("127.0.0.1:53802")
 	assert.Nil(t, err)
 
-	localP2P := newP2p(43802, "testdata2", q)
+	localP2P := newP2p(43802, "testPeer", q)
 	defer freeP2p(localP2P)
 
 	t.Log(localP2P.node.CacheBoundsSize())
@@ -222,7 +209,7 @@ func TestPeer(t *testing.T) {
 	p2pcli := NewNormalP2PCli()
 	localP2P.node.nodeInfo.peerInfos.SetPeerInfo(nil)
 	localP2P.node.nodeInfo.peerInfos.GetPeerInfo("1222")
-	t.Log(p2pModule.node.GetRegisterPeer("localhost:43802"))
+	t.Log(p2p.node.GetRegisterPeer("localhost:43802"))
 	//测试发送Ping消息
 	err = p2pcli.SendPing(peer, localP2P.node.nodeInfo)
 	assert.Nil(t, err)
@@ -273,23 +260,10 @@ func TestPeer(t *testing.T) {
 	job.setFreePeer(peer.GetPeerName())
 	job.removePeer(peer.GetPeerName())
 	job.CancelJob()
-	os.Remove(dataDir)
-
-}
-
-func TestSortArr(t *testing.T) {
-	var Inventorys = make(Invs, 0)
-	for i := 100; i >= 0; i-- {
-		var inv types.Inventory
-		inv.Ty = 111
-		inv.Height = int64(i)
-		Inventorys = append(Inventorys, &inv)
-	}
-	sort.Sort(Inventorys)
 }
 
 //测试grpc 多连接
-func TestGrpcConns(t *testing.T) {
+func testGrpcConns(t *testing.T) {
 	var conns []*grpc.ClientConn
 
 	for i := 0; i < maxSamIPNum; i++ {
@@ -320,7 +294,7 @@ func TestGrpcConns(t *testing.T) {
 }
 
 //测试grpc 流多连接
-func TestGrpcStreamConns(t *testing.T) {
+func testGrpcStreamConns(t *testing.T, p2p *P2p) {
 
 	conn, err := grpc.Dial("localhost:53802", grpc.WithInsecure(),
 		grpc.WithDefaultCallOptions(grpc.UseCompressor("gzip")))
@@ -332,7 +306,7 @@ func TestGrpcStreamConns(t *testing.T) {
 	_, err = resp.Recv()
 	assert.Equal(t, true, strings.Contains(err.Error(), "no authorized"))
 
-	ping, err := P2pComm.NewPingData(p2pModule.node.nodeInfo)
+	ping, err := P2pComm.NewPingData(p2p.node.nodeInfo)
 	assert.Nil(t, err)
 
 	_, err = cli.ServerStreamSend(context.Background(), ping)
@@ -349,54 +323,22 @@ func TestGrpcStreamConns(t *testing.T) {
 
 }
 
-func TestP2pComm(t *testing.T) {
+func testP2pComm(t *testing.T, p2p *P2p) {
 
 	addrs := P2pComm.AddrRouteble([]string{"localhost:53802"}, calcChannelVersion(testChannel))
 	t.Log(addrs)
-
 	i32 := P2pComm.BytesToInt32([]byte{0xff})
 	t.Log(i32)
-
 	_, _, err := P2pComm.GenPrivPubkey()
 	assert.Nil(t, err)
-
-	ping, err := P2pComm.NewPingData(p2pModule.node.nodeInfo)
+	ping, err := P2pComm.NewPingData(p2p.node.nodeInfo)
 	assert.Nil(t, err)
-
 	assert.Equal(t, true, P2pComm.CheckSign(ping))
 	assert.IsType(t, "string", P2pComm.GetLocalAddr())
 	assert.Equal(t, 5, len(P2pComm.RandStr(5)))
-
 }
 
-func TestFilter(t *testing.T) {
-	filter := NewFilter(10)
-	go filter.ManageRecvFilter()
-	defer filter.Close()
-	filter.GetLock()
-
-	assert.Equal(t, true, filter.RegRecvData("key"))
-	assert.Equal(t, true, filter.QueryRecvData("key"))
-	filter.RemoveRecvData("key")
-	assert.Equal(t, false, filter.QueryRecvData("key"))
-	filter.ReleaseLock()
-
-}
-
-func TestAddrRouteble(t *testing.T) {
-	resp := P2pComm.AddrRouteble([]string{"114.55.101.159:13802"}, calcChannelVersion(testChannel))
-	t.Log(resp)
-}
-
-func TestRandStr(t *testing.T) {
-	t.Log(P2pComm.RandStr(5))
-}
-
-func TestGetLocalAddr(t *testing.T) {
-	t.Log(P2pComm.GetLocalAddr())
-}
-
-func TestAddrBook(t *testing.T) {
+func testAddrBook(t *testing.T, p2p *P2p) {
 
 	prv, pub, err := P2pComm.GenPrivPubkey()
 	if err != nil {
@@ -413,7 +355,7 @@ func TestAddrBook(t *testing.T) {
 	}
 	t.Log("GenPubkey:", pubstr)
 
-	addrBook := p2pModule.node.nodeInfo.addrBook
+	addrBook := p2p.node.nodeInfo.addrBook
 	addrBook.Size()
 	addrBook.saveToDb()
 	addrBook.GetPeerStat("locolhost:43802")
@@ -427,217 +369,34 @@ func TestAddrBook(t *testing.T) {
 	addrBook.ResetPeerkey(hex.EncodeToString(prv), pubstr)
 	resetkey, _ := addrBook.GetPrivPubKey()
 	assert.NotEqual(t, resetkey, privkey)
-
 }
 
-func TestBytesToInt32(t *testing.T) {
-
-	t.Log(P2pComm.BytesToInt32([]byte{0xff}))
-	t.Log(P2pComm.Int32ToBytes(255))
-}
-
-func TestNetAddress(t *testing.T) {
-	tcpAddr := new(net.TCPAddr)
-	tcpAddr.IP = net.ParseIP("localhost")
-	tcpAddr.Port = 2223
-	nad := NewNetAddress(tcpAddr)
-	nad1 := nad.Copy()
-	nad.Equals(nad1)
-	nad2s, err := NewNetAddressStrings([]string{"localhost:3306"})
-	if err != nil {
-		return
-	}
-	nad.Less(nad2s[0])
-
-}
-
-func TestP2pListen(t *testing.T) {
-	var node Node
-	node.listenPort = 3333
-	listen1 := NewListener("tcp", &node)
-	assert.Equal(t, true, listen1 != nil)
-	listen2 := NewListener("tcp", &node)
-	assert.Equal(t, true, listen2 != nil)
-
-	listen1.Close()
-	listen2.Close()
-}
-
-func TestP2pRestart(t *testing.T) {
-
-	p2p := newP2p(12345, "testRestart", queue.New("channel"))
-	defer freeP2p(p2p)
-	assert.Equal(t, false, p2p.isClose())
-	assert.Equal(t, false, p2p.isRestart())
-	for i := 0; i < 100; i++ {
-		client := p2p.client
-		msg := client.NewMessage("p2p", types.EventTxBroadcast, &types.Transaction{})
-		assert.Nil(t, client.Send(msg, false))
-	}
+func testRestart(t *testing.T, p2p *P2p) {
+	client := p2p.client
+	assert.False(t, p2p.isRestart())
+	p2p.txFactory <- struct{}{}
+	p2p.processEvent(client.NewMessage("p2p", types.EventTxBroadcast, &types.Transaction{}), 128, p2p.p2pCli.BroadCastTx)
+	atomic.StoreInt32(&p2p.restart, 1)
+	p2p.ReStart()
+	atomic.StoreInt32(&p2p.restart, 0)
 	p2p.ReStart()
 }
 
-type versionData struct {
-	rawData interface{}
-	version int32
-}
-
-func Test_processP2P(t *testing.T) {
+func Test_p2p(t *testing.T) {
 
 	q := queue.New("channel")
-	p2p := newP2p(12345, "testProcessP2p", q)
+	go q.Start()
+	processMsg(q)
+	p2p := newP2p(53802, "testP2p", q)
+	p2p.Wait()
 	defer freeP2p(p2p)
-	node := p2p.node
-	client := p2p.client
-	pid := "testPid"
-	sendChan := make(chan interface{}, 1)
-	recvChan := make(chan *types.BroadCastData, 1)
-	testDone := make(chan struct{})
-
-	payload := []byte("testpayload")
-	minerTx := &types.Transaction{Execer: []byte("coins"), Payload: payload, Fee: 14600, Expire: 200}
-	tx := &types.Transaction{Execer: []byte("coins"), Payload: payload, Fee: 4600, Expire: 2}
-	tx1 := &types.Transaction{Execer: []byte("coins"), Payload: payload, Fee: 460000000, Expire: 0}
-	tx2 := &types.Transaction{Execer: []byte("coins"), Payload: payload, Fee: 100, Expire: 1}
-	txGroup, _ := types.CreateTxGroup([]*types.Transaction{tx1, tx2})
-	gtx := txGroup.Tx()
-	txList := append([]*types.Transaction{}, minerTx, tx, tx1, tx2)
-	memTxList := append([]*types.Transaction{}, tx, gtx)
-
-	block := &types.Block{
-		TxHash: []byte("123"),
-		Height: 10,
-		Txs:    txList,
-	}
-	txHash := hex.EncodeToString(tx.Hash())
-	blockHash := hex.EncodeToString(block.Hash())
-	rootHash := merkle.CalcMerkleRoot(txList)
-
-	//mempool handler
-	go func() {
-		client := q.Client()
-		client.Sub("mempool")
-		for msg := range client.Recv() {
-			switch msg.Ty {
-			case types.EventTxListByHash:
-				query := msg.Data.(*types.ReqTxHashList)
-				var txs []*types.Transaction
-				if !query.IsShortHash {
-					txs = memTxList[:1]
-				} else {
-					txs = memTxList
-				}
-				msg.Reply(client.NewMessage("p2p", types.EventTxListByHash, &types.ReplyTxList{Txs: txs}))
-			}
-		}
-	}()
-
-	//测试发送
-	go func() {
-		for data := range sendChan {
-			verData, ok := data.(*versionData)
-			assert.True(t, ok)
-			sendData, doSend := node.processSendP2P(verData.rawData, verData.version, "testIP:port")
-			txHashFilter.regRData.Remove(txHash)
-			blockHashFilter.regRData.Remove(blockHash)
-			assert.True(t, doSend, "sendData:", verData)
-			recvChan <- sendData
-		}
-	}()
-	//测试接收
-	go func() {
-		for data := range recvChan {
-			txHashFilter.regRData.Remove(txHash)
-			blockHashFilter.regRData.Remove(blockHash)
-			handled := node.processRecvP2P(data, pid, node.pubToPeer, "testIP:port")
-			assert.True(t, handled)
-		}
-	}()
-
-	go func() {
-		p2pChan := node.pubsub.Sub("tx")
-		for data := range p2pChan {
-			if p2pTx, ok := data.(*types.P2PTx); ok {
-				sendChan <- &versionData{rawData: p2pTx, version: lightBroadCastVersion}
-			}
-		}
-	}()
-
-	//data test
-	go func() {
-		subChan := node.pubsub.Sub(pid)
-		//normal
-		sendChan <- &versionData{rawData: &types.P2PTx{Tx: tx, Route: &types.P2PRoute{}}, version: lightBroadCastVersion - 1}
-		assert.Nil(t, client.Send(client.NewMessage("p2p", types.EventTxBroadcast, tx), false))
-		sendChan <- &versionData{rawData: &types.P2PBlock{Block: block}, version: lightBroadCastVersion - 1}
-		//light broadcast
-		txHashFilter.Add(hex.EncodeToString(tx1.Hash()), &types.P2PRoute{TTL: DefaultLtTxBroadCastTTL})
-		_ = client.Send(client.NewMessage("p2p", types.EventTxBroadcast, tx1), false)
-		sendChan <- &versionData{rawData: &types.P2PTx{Tx: tx, Route: &types.P2PRoute{TTL: DefaultLtTxBroadCastTTL}}, version: lightBroadCastVersion}
-		<-subChan //query tx
-		sendChan <- &versionData{rawData: &types.P2PBlock{Block: block}, version: lightBroadCastVersion}
-		<-subChan //query block
-		for !ltBlockCache.contains(blockHash) {
-		}
-		ltBlock := ltBlockCache.get(blockHash).(*types.Block)
-		assert.True(t, bytes.Equal(rootHash, merkle.CalcMerkleRoot(ltBlock.Txs)))
-
-		//query tx
-		sendChan <- &versionData{rawData: &types.P2PQueryData{Value: &types.P2PQueryData_TxReq{TxReq: &types.P2PTxReq{TxHash: tx.Hash()}}}}
-		_, ok := (<-subChan).(*types.P2PTx)
-		assert.True(t, ok)
-		sendChan <- &versionData{rawData: &types.P2PQueryData{Value: &types.P2PQueryData_BlockTxReq{BlockTxReq: &types.P2PBlockTxReq{
-			BlockHash: blockHash,
-			TxIndices: []int32{1, 2},
-		}}}}
-		rep, ok := (<-subChan).(*types.P2PBlockTxReply)
-		assert.True(t, ok)
-		assert.Equal(t, 2, int(rep.TxIndices[1]))
-		sendChan <- &versionData{rawData: &types.P2PQueryData{Value: &types.P2PQueryData_BlockTxReq{BlockTxReq: &types.P2PBlockTxReq{
-			BlockHash: blockHash,
-			TxIndices: nil,
-		}}}}
-		rep, ok = (<-subChan).(*types.P2PBlockTxReply)
-		assert.True(t, ok)
-		assert.Nil(t, rep.TxIndices)
-
-		//query reply
-		sendChan <- &versionData{rawData: &types.P2PBlockTxReply{
-			BlockHash: blockHash,
-			TxIndices: []int32{1},
-			Txs:       txList[1:2],
-		}}
-		<-subChan
-		assert.True(t, ltBlockCache.contains(blockHash))
-		assert.Nil(t, ltBlock.Txs)
-
-		ltBlock.TxHash = rootHash
-		sendChan <- &versionData{rawData: &types.P2PBlockTxReply{
-			BlockHash: blockHash,
-			Txs:       txList[0:],
-		}}
-		for ltBlockCache.contains(blockHash) {
-		}
-		//max ttl
-		node.nodeInfo.cfg.MaxTTL = 0
-		_, doSend := node.processSendP2P(&types.P2PTx{Tx: tx, Route: &types.P2PRoute{TTL: 1}}, lightBroadCastVersion, "testIP:port")
-		assert.False(t, doSend)
-		close(testDone)
-	}()
-
-	ticker := time.NewTicker(time.Minute)
-	defer ticker.Stop()
-	for {
-		select {
-		case <-testDone:
-			return
-		case <-ticker.C:
-			t.Error("TestP2PProcessTimeout")
-			return
-		}
-	}
-}
-
-func TestP2pClose(t *testing.T) {
-	freeP2p(p2pModule)
+	defer q.Close()
+	testP2PEvent(t, q.Client())
+	testNetInfo(t, p2p)
+	testPeer(t, p2p, q)
+	testGrpcConns(t)
+	testGrpcStreamConns(t, p2p)
+	testP2pComm(t, p2p)
+	testAddrBook(t, p2p)
+	testRestart(t, p2p)
 }
